2023/05/23 Update) Application.Quit() Crash 해결
Update)
-Application.Quit시 Json 데이터로 저장하는 것과 crash가 나던 이슈 해결

+++ Assets/Scripts/Managers/GameManager.cs
@@ -238,7 +238,6 @@ public class GameManager : Singleton<GameManager>, IDisposable
     {
         //테스트 버전에선 무조건 3개다 열어두기
         AudioManager.GetInstance.PlaySFX(AudioManager.SFX.UIClick);
-        DataManager.GetInstance.SaveData(3);
 #if UNITY_EDITOR
         EditorApplication.isPlaying = false;
 #else
@@ -248,8 +247,9 @@ public class GameManager : Singleton<GameManager>, IDisposable
 
     private void OnApplicationQuit()
     {
-#if !UNITY_EDITOR
-        System.Diagnostics.Process.GetCurrentProcess().Kill();
-#endif
+        if (!Application.isEditor)
+        {
+            System.Diagnostics.Process.GetCurrentProcess().Kill();
+        }
     }
 }
